♻️ refactor: handleSlotChat을 SlotChat컴포넌트의 export로 변경
